Removed all comments from the "configure.ac" file. This makes it easier to read, and you can still look up the comments in the git history / Google.
Consolidated all "Makefile.am" files into a single file. This means all
GNU Autotools configuration is now done in just two files, both in the
root directory of the project: "configure.ac" and "Makefile.am".

Re-added the "PKGDATADIR", but with a twist: if the variable already
exists (meaning, you are compiling with GNU Autotools) then it's good to
go. If it DOESN'T exist (for example, when compiling from an IDE or
using "gcc" directly) it will simply set the variable to the local
"../data" directory. See "Resources.h" for more information.